Move PublicationRegisterRequest to schemas
